pyproject.toml
[tool.poetry]
name = "mysraper"
version = "0.1.0"
description = ""
authors = ["celedituro <cdituro@fi.uba.ar>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.15.0"
pytest = "^6.2.5"
fastapi = "^0.68.1"
requests = "^2.31.0"
psycopg2 = "^2.9.1"
psycopg2-binary = "^2.9.3"
python-dotenv = "^0.21.0"
telethon = "^1.29.2"
httpx = "^0.24.1"
pydantic = "^1.8.2"
bcrypt = "^4.0.1"
pyjwt = "^2.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore:.*'imghdr' is deprecated.*:DeprecationWarning"
]